Use Symbols for enums.
haphazard tests show a possible 5-10% improvement on Adventure.
originally we went for strings for simplicity and because simple benches
showed little difference between strings and other things (and this may
well be the case) but at the least it doesn't seem harmful to use
Symbols (classic two-way JS enums may be just as good), and the use of
extensive switch statements in a loop is not something ive done bench
comparisons of. it seems so silly and superstitious but thats the game i
guess